Move DonationsOrClaims to library/
Presently, significant code representing instantiable components,
"Donation"s and "Claim"s, are defined inside "DashboardScreen/". These
components may be required across screens such as donations-history
or claims-history pages. While putting the code inside a shareable
directory inside "screens/" is also an option, putting them inside a
currently-non-existent "library/" directory which would represent
concrete components is a better option. This follows the choice to host
common page components inside "elements/" as opposed to putting them
inside a shareable directory within "screens/".
